+++ apps/element-storybook/.storybook/preview.tsx
@@ -71,4 +71,19 @@ const preview: Preview = {
   },
 };
 
+// Make sure we are in the browser before starting
+if (typeof global.process === 'undefined') {
+  // eslint-disable-next-line @nrwl/nx/enforce-module-boundaries
+  import('../../../packages/mock/src/lib/browser').then(({ worker }) => {
+    const config =
+      process.env.NODE_ENV === 'development'
+        ? undefined
+        : {
+            serviceWorker: { url: '/element/mockServiceWorker.js' },
+            onUnhandledRequest: 'bypass',
+          };
+    worker.start(config);
+  });
+}
+
 export default preview;