+++ packages/autocomplete/src/lib/Autocomplete.stories.tsx
@@ -1,20 +1,7 @@
 // Each exported component in the package should have its own stories file
 
 import type { Meta, StoryObj } from '@storybook/react';
-import { Autocomplete, AutocompleteProps } from './Autocomplete';
-import {
-  Box,
-  Chip,
-  MenuItem,
-  Autocomplete as MuiAutocomplete,
-  AutocompleteProps as MuiAutocompleteProps,
-  TextField as MuiTextField,
-  SelectChangeEvent,
-  Typography,
-} from '@mui/material';
-import { TextField } from '@availity/mui-textfield';
-import { Select, SelectPropOverrides } from '@availity/mui-form-utils';
-import { useState } from 'react';
+import { Autocomplete } from './Autocomplete';
 
 const meta: Meta<typeof Autocomplete> = {
   title: 'Components/Autocomplete/Autocomplete',
@@ -23,6 +10,13 @@ const meta: Meta<typeof Autocomplete> = {
   args: {
     options: ['1', '2', '3', '4', '5'],
   },
+  argTypes: {
+    multiple: {
+      table: {
+        disable: true,
+      },
+    },
+  },
 };
 
 export default meta;
@@ -30,15 +24,21 @@ export default meta;
 export const _Autocomplete: StoryObj<typeof Autocomplete> = {
   render: (args) => <Autocomplete {...args} />,
   args: {
-    FieldProps: { label: 'AvTextField' },
+    FieldProps: { label: 'Autocomplete', helperText: 'Helper Text', fullWidth: false },
   },
 };
 
-export const _Variants: StoryObj<typeof Autocomplete> = {
-  render: (args) => (
-    <>
-      <Autocomplete {...args} FieldProps={{ label: 'Single Select', margin: 'normal' }} />
-      <Autocomplete {...args} FieldProps={{ label: 'Multi Select', margin: 'normal' }} multiple />
-    </>
-  ),
+export const _Single: StoryObj<typeof Autocomplete> = {
+  render: (args) => <Autocomplete {...args} />,
+  args: {
+    FieldProps: { label: 'Single Select', helperText: 'Helper Text', fullWidth: false },
+  },
+};
+
+export const _Multi: StoryObj<typeof Autocomplete> = {
+  render: (args) => <Autocomplete {...args} />,
+  args: {
+    FieldProps: { label: 'Multi Select', helperText: 'Helper Text', fullWidth: false },
+    multiple: true,
+  },
 };