+++ packages/theme/src/lib/legacy-theme.ts
@@ -13,7 +13,9 @@ const containedButtonStyles = (color: string) => ({
   '&:focus': {
     backgroundColor: tokens[`color${color}Dark` as keyof typeof tokens],
     outline: '2px solid white',
-    boxShadow: `0 0 0px 4px ${tokens[`color${color}Dark` as keyof typeof tokens]}`,
+    boxShadow: `0 0 0px 4px ${
+      color === 'Tertiary' ? tokens.colorSecondaryDark : tokens[`color${color}Dark` as keyof typeof tokens]
+    }`,
     color: color === 'Warning' && tokens.colorTextInverse,
   },
   '&:active': {
@@ -30,7 +32,9 @@ const outlinedButtonStyles = (color: string) => ({
   },
   '&:focus': {
     outline: '2px solid white',
-    boxShadow: `0 0 0px 4px ${tokens[`color${color}Main` as keyof typeof tokens]}`,
+    boxShadow: `0 0 0px 4px ${
+      color === 'Tertiary' ? tokens.colorSecondaryMain : tokens[`color${color}Main` as keyof typeof tokens]
+    }`,
     backgroundColor: `${tokens[`color${color}Main` as keyof typeof tokens]}21`,
   },
   '&:active': {
@@ -174,11 +178,13 @@ export const legacyTheme = {
       hover: tokens.colorActionHover,
       hoverOpacity: 0.04,
       selected: tokens.colorActionSelected,
+      // selected: 'rgb(50, 98, 175)',
       selectedOpacity: 0.08,
       disabled: tokens.colorActionDisabled,
       disabledBackground: tokens.colorActionDisabledBg,
       disabledOpacity: 0.38,
       focus: tokens.colorActionFocus,
+      // focus: 'rgb(184, 212, 251)',
       focusOpacity: 0.12,
       activatedOpacity: 0.12,
     },
@@ -299,6 +305,50 @@ export const legacyTheme = {
     },
   },
   components: {
+    AvFeedbackContainer: {
+      styleOverrides: {
+        root: {
+          maxWidth: important('490px'),
+          padding: '16px',
+        },
+        smileButtons: {
+          marginBottom: '8px',
+          '.MuiToggleButton-root': {
+            margin: '8px',
+            '&:focus': {
+              backgroundColor: tokens.colorActionFocus,
+              color: tokens.colorTextPrimary,
+              boxShadow: 'none',
+            },
+            '&:hover': {
+              backgroundColor: tokens.colorActionHover,
+              color: tokens.colorTextPrimary,
+              borderColor: tokens.colorGrey400,
+            },
+            '&.Mui-selected': {
+              backgroundColor: tokens.colorPrimaryMain,
+              '&:active': {
+                backgroundColor: tokens.colorPrimaryDarker,
+              },
+              '&:focus': {
+                boxShadow: 'none',
+              },
+              '&:focus, &:hover': {
+                backgroundColor: tokens.colorPrimaryDark,
+                color: tokens.colorCommonWhite,
+              },
+              '&.Mui-disabled': {
+                backgroundColor: tokens.colorPrimaryMain20,
+              },
+            },
+            '&.MuiToggleButton-sizeMedium': {
+              padding: '12px',
+              height: 'auto',
+            },
+          },
+        },
+      },
+    },
     MuiAlert: {
       styleOverrides: {
         standardSuccess: {
@@ -360,6 +410,81 @@ export const legacyTheme = {
         },
       },
     },
+    MuiAutocomplete: {
+      defaultProps: {
+        size: 'small',
+        fullWidth: true,
+        clearOnEscape: true,
+        ChipProps: {
+          size: 'medium',
+        },
+      },
+      styleOverrides: {
+        root: {
+          '&.Mui-error .MuiAutocomplete-popupIndicator': {
+            color: tokens.borderError,
+          },
+          '&.MuiAutocomplete-hasPopupIcon': {
+            '.MuiAutocomplete-endAdornment': {
+              right: '0px',
+            },
+            '.MuiInputBase-root.MuiInputBase-adornedEnd': {
+              paddingRight: '2.5rem',
+            },
+            '&.MuiAutocomplete-hasClearIcon': {
+              '.MuiInputBase-root.MuiInputBase-adornedEnd': {
+                paddingRight: '4rem',
+              },
+            },
+          },
+        },
+        endAdornment: {
+          top: 'unset',
+          right: '4px',
+          height: 'calc(100% - 16px)',
+          display: 'inline-flex',
+          alignItems: 'center',
+          '.MuiIconButton-root': {
+            display: 'flex',
+            marginRight: '0',
+            width: '2.25rem',
+            height: '2.25rem',
+            '.MuiSelect-avExpandIcon': {
+              top: 'unset',
+              transform: 'unset',
+            },
+          },
+        },
+        popupIndicator: {
+          color: tokens.borderInput,
+        },
+        popupIndicatorOpen: {
+          transform: 'scaleY(-1)',
+        },
+        clearIndicator: {
+          color: tokens.borderInput,
+          marginTop: '-8px',
+          marginBottom: '-8px',
+        },
+        option: {
+          '&.MuiAutocomplete-option': {
+            minHeight: 'unset',
+            "&[aria-selected='true'], &[aria-selected='true'].Mui-focused": {
+              color: tokens.colorCommonWhite,
+              backgroundColor: 'rgba(50, 98, 175, 1)',
+              fontWeight: 500,
+            },
+            '&.Mui-focused': {
+              color: tokens.colorTextPrimary,
+              backgroundColor: 'rgba(184, 212, 251,1)',
+            },
+          },
+        },
+        tag: {
+          margin: '2px',
+        },
+      },
+    },
     MuiAvatar: {
       styleOverrides: {
         colorDefault: {
@@ -410,14 +535,7 @@ export const legacyTheme = {
         },
         containedPrimary: { ...containedButtonStyles('Primary') },
         containedSecondary: { ...containedButtonStyles('Secondary') },
-        containedTertiary: {
-          ...containedButtonStyles('Tertiary'),
-          '&:focus': {
-            backgroundColor: tokens.colorTertiaryDark,
-            outline: '2px solid white',
-            boxShadow: `0 0 0px 4px ${tokens.colorSecondaryDark}`,
-          },
-        },
+        containedTertiary: { ...containedButtonStyles('Tertiary') },
         containedSuccess: { ...containedButtonStyles('Success') },
         containedWarning: { ...containedButtonStyles('Warning') },
         containedError: { ...containedButtonStyles('Error') },
@@ -635,6 +753,12 @@ export const legacyTheme = {
         },
       },
     },
+    MuiFormControl: {
+      defaultProps: {
+        size: 'small',
+        fullWidth: true,
+      },
+    },
     MuiFormHelperText: {
       styleOverrides: {
         root: {
@@ -656,10 +780,7 @@ export const legacyTheme = {
     MuiFormLabel: {
       styleOverrides: {
         root: {
-          // move required asterisk before text
-          display: important('flex'),
-          flexDirection: important('row-reverse'),
-          justifyContent: important('flex-end'),
+          marginBottom: '.1rem',
           '&.Mui-error, &.Mui-error.Mui-focused': {
             color: tokens.colorTextError,
           },
@@ -669,6 +790,10 @@ export const legacyTheme = {
           '&.Mui-focused': {
             color: 'inherit',
           },
+          // move required asterisk before text
+          display: important('flex'),
+          flexDirection: important('row-reverse'),
+          justifyContent: important('flex-end'),
         },
         asterisk: {
           color: tokens.colorErrorMain,
@@ -677,9 +802,12 @@ export const legacyTheme = {
           fontSize: '1.3rem',
           lineHeight: '100%',
         },
-        children: {
-          '.MuiFormLabel-asterisk': {
-            display: 'none',
+        wrapper: {
+          marginBottom: '.1rem',
+          display: 'flex',
+          flexDirection: 'row',
+          '.MuiFormLabel-root': {
+            marginBottom: '.1rem',
           },
         },
       },
@@ -730,16 +858,33 @@ export const legacyTheme = {
         size: 'small',
       },
       styleOverrides: {
-        root: {
-          'label + &': {
-            marginTop: '.5rem',
-          },
-          borderRadius: 4,
-          position: 'relative',
-          fontSize: '1rem',
-        },
-        inputSizeSmall: {
-          padding: '.375rem .75rem',
+        root: ({ ownerState }: { ownerState: { fullWidth: boolean | undefined } }) => {
+          return {
+            boxSizing: 'border-box',
+            '&.MuiInputBase-root, &.MuiInputBase-root.MuiAutocomplete-inputRoot': {
+              width: ownerState.fullWidth ? '100%' : '220px',
+            },
+            'label + &': {
+              marginTop: '.5rem',
+            },
+            borderRadius: 4,
+            position: 'relative',
+            fontSize: '1rem',
+            '.MuiOutlinedInput-input': {
+              '&.MuiInputBase-inputSizeSmall': {
+                padding: '.375rem .75rem',
+                lineHeight: '24px',
+              },
+            },
+            '.MuiSelect-avEndAdornmentDivider': {
+              borderColor: tokens.colorBackgroundBorder,
+            },
+            '.MuiChip-root': {
+              lineHeight: '1.125rem',
+              fontSize: '.875rem',
+              padding: '3px 6px',
+            },
+          };
         },
       },
     },
@@ -886,8 +1031,27 @@ export const legacyTheme = {
             color: tokens.borderError,
           },
         },
+        avExpandIcon: {
+          position: 'relative',
+          top: '50%',
+          transform: 'translateY(-50%)',
+        },
         icon: {
           color: tokens.borderInput,
+          '&.MuiSvgIcon-root:not(.MuiTablePagination-selectIcon)': {
+            width: '2rem',
+            height: '2rem',
+            top: 'unset',
+          },
+          '&.MuiStack-root': {
+            right: '4px',
+            height: '100%',
+            padding: '8px 0px',
+            '.MuiSvgIcon-root': {
+              width: '2rem',
+              paddingLeft: '4px',
+            },
+          },
         },
         iconOpen: {
           transform: 'scaleY(-1)',
@@ -1013,6 +1177,9 @@ export const legacyTheme = {
           border: `1px solid ${tokens.borderInput}`,
           marginLeft: '.25rem',
           marginRight: '.75rem',
+          '&.MuiInputBase-root': {
+            width: 'auto',
+          },
         },
         select: {
           '&.MuiTablePagination-select.MuiSelect-select': {
@@ -1070,11 +1237,6 @@ export const legacyTheme = {
           backgroundColor: tokens.colorBackgroundCanvas,
           height: '3rem',
         },
-      },
-    },
-    MuiTextField: {
-      defaultProps: {
-        fullWidth: true,
       },
     },
     MuiToggleButton: {