+++ packages/form-utils/CHANGELOG.md
@@ -2,17 +2,29 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.6.1](https://github.com/Availity/element/compare/@availity/mui-form-utils@0.6.0...@availity/mui-form-utils@0.6.1) (2023-11-27)
+
+### Dependency Updates
+
+- `mui-chip` updated to version `0.2.5`
+
+## [0.6.0](https://github.com/Availity/element/compare/@availity/mui-form-utils@0.5.4...@availity/mui-form-utils@0.6.0) (2023-11-22)
+
+### Features
+
+- **mui-form-utils:** new slots and handling of required ([332f1e8](https://github.com/Availity/element/commit/332f1e8fa18d9c2d5492fd6b9e315e3ccc5fa7e9))
+
 ## [0.5.4](https://github.com/Availity/element/compare/@availity/mui-form-utils@0.5.3...@availity/mui-form-utils@0.5.4) (2023-11-16)
 
 ### Dependency Updates
 
-* `mui-button` updated to version `0.5.1`
-* `mui-icon` updated to version `0.7.3`
-* `mui-chip` updated to version `0.2.4`
+- `mui-button` updated to version `0.5.1`
+- `mui-icon` updated to version `0.7.3`
+- `mui-chip` updated to version `0.2.4`
 
 ### Bug Fixes
 
-* **mui-form-utils:** add back sx props ([39f5cb2](https://github.com/Availity/element/commit/39f5cb208d835750aa83f40e685841ff244485d8))
+- **mui-form-utils:** add back sx props ([39f5cb2](https://github.com/Availity/element/commit/39f5cb208d835750aa83f40e685841ff244485d8))
 
 ## [0.5.3](https://github.com/Availity/element/compare/@availity/mui-form-utils@0.5.2...@availity/mui-form-utils@0.5.3) (2023-11-08)