+++ packages/menu/CHANGELOG.md
@@ -6,8 +6,9 @@ This file was generated using [@jscutlery/semver](https://github.com/jscutlery/s
 
 ### Dependency Updates
 
-* `mui-icon` updated to version `0.7.3`
-* `mui-button` updated to version `0.5.1`
+- `mui-icon` updated to version `0.7.3`
+- `mui-button` updated to version `0.5.1`
+
 ## [0.1.6](https://github.com/Availity/element/compare/@availity/mui-menu@0.1.5...@availity/mui-menu@0.1.6) (2023-11-08)
 
 ### Dependency Updates