+++ packages/progress/CHANGELOG.md
@@ -6,8 +6,9 @@ This file was generated using [@jscutlery/semver](https://github.com/jscutlery/s
 
 ### Dependency Updates
 
-* `mui-icon` updated to version `0.7.3`
-* `mui-typography` updated to version `0.1.5`
+- `mui-icon` updated to version `0.7.3`
+- `mui-typography` updated to version `0.1.5`
+
 ## [0.1.1](https://github.com/Availity/element/compare/@availity/mui-progress@0.1.0...@availity/mui-progress@0.1.1) (2023-11-15)
 
 ## 0.1.0 (2023-11-14)