+++ packages/table/CHANGELOG.md
@@ -2,9 +2,12 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
-## 0.1.0 (2023-11-16)
+## [0.1.2](https://github.com/Availity/element/compare/@availity/mui-table@0.1.1...@availity/mui-table@0.1.2) (2023-11-28)
 
+## [0.1.1](https://github.com/Availity/element/compare/@availity/mui-table@0.1.0...@availity/mui-table@0.1.1) (2023-11-27)
+
+## 0.1.0 (2023-11-16)
 
 ### Features
 
-* **mui-table:** add table components ([53dfe7e](https://github.com/Availity/element/commit/53dfe7e7d6282cb36a12b7114597d1b230d1ccdb))
+- **mui-table:** add table components ([53dfe7e](https://github.com/Availity/element/commit/53dfe7e7d6282cb36a12b7114597d1b230d1ccdb))